adjustments

+++ data/src/main/java/com/fioalpha/poc/data/networking/model/RetrofitService.kt
@@ -7,7 +7,7 @@ import retrofit2.http.Query
 interface RetrofitService {
     @GET("/v3/ecfaebf5-782b-4b24-ae4f-23b5c3a861da")
     suspend fun fetchInvestment(
-        @Query("investedAmount") investedAmount: Int,
+        @Query("investedAmount") investedAmount: Double,
         @Query("maturityDate") maturityDate: String,
         @Query("rate") rate: Int,
         @Query("isTaxFree") isTaxFree: Boolean = false,
